Consolidate pathway expression code, denoise errors
