+++ app/javascript/components/visualization/DotPlot.jsx
@@ -180,118 +180,6 @@ function patchServiceWorkerCache() {
   }
 }
 
-/**
- * Get mean, percent, and color per gene, by annotation label
- *
- * For scRNA-seq:
- * - Mean: average expression value
- * - Percent: percent of cells expressing
- * - Color: hex value for color of scaled mean expression: blue low, purple medium, red high
- */
-export function getDotPlotMetrics(dotPlot) {
-  const metrics = {}
-
-  const colorScheme = dotPlot.getColorScheme()
-
-  const dataset = dotPlot.dataset
-  const labels = dataset.columnMetadata.vectors[0].array
-  const genes = dataset.rowMetadata.vectors[0].array
-
-  for (let i = 0; i < labels.length; i++) {
-    const label = labels[i]
-    const labelIndex = i
-    metrics[label] = {}
-    for (let j = 0; j < genes.length; j++) {
-      const gene = genes[j]
-      const geneIndex = j
-      try {
-        const mean = dataset.getValue(geneIndex, labelIndex, 0)
-        const percent = dataset.getValue(geneIndex, labelIndex, 1)
-        const color = colorScheme.getColor(geneIndex, labelIndex, mean)
-        metrics[label][gene] = { mean, percent, color }
-      } catch (error) {
-        // eslint-disable-next-line quotes
-        if (error.message === "Cannot read properties of undefined (reading 'getValue')") {
-          // Occurs upon resizing window, artifact of internal Morpheus handling
-          // of pre-dot-plot heatmap matrix.  No user-facing impact.
-          return null
-        }
-      }
-    }
-  }
-
-  return metrics
-}
-
-window.SCP.renderBackgroundDotPlotRegister = {}
-
-/** Render undisplayed Morpheus dot plot, to get metrics for pathway diagram */
-export async function renderBackgroundDotPlot(
-  studyAccession, genes=[], cluster, annotation={},
-  subsample, annotationValues, drawCallback,
-  topContainerSelector
-) {
-  const graphId = 'background-dot-plot'
-  document.querySelector(`#${graphId}`)?.remove()
-
-  const registerKey = [
-    studyAccession,
-    genes.join(','),
-    cluster,
-    annotation.name,
-    annotation.type,
-    annotation.scope,
-    subsample
-  ].join('--')
-
-  // Prevent duplicate parallel requests of the same dot plot
-  if (registerKey in window.SCP.renderBackgroundDotPlotRegister) {
-    console.log('Prevented duplicate parallel requests of background dot plot')
-    return
-  }
-  window.SCP.renderBackgroundDotPlotRegister[registerKey] = 1
-
-  const topContainer = document.querySelector(topContainerSelector)
-
-  const container = `<div id="${graphId}" style="display: none;">`
-
-  topContainer.insertAdjacentHTML('beforeEnd', container)
-  const target = `#${graphId}`
-
-  performance.mark(`perfTimeStart-${graphId}`)
-  let dataset
-
-  try {
-    const results = await fetchMorpheusJson(
-      studyAccession,
-      genes,
-      cluster,
-      annotation.name,
-      annotation.type,
-      annotation.scope,
-      subsample
-    )
-    dataset = results[0]
-
-    // Don't prevent non-parallel duplicate requests
-    delete window.SCP.renderBackgroundDotPlotRegister[registerKey]
-  } catch (error) {
-    delete window.SCP.renderBackgroundDotPlotRegister[registerKey]
-  }
-
-
-  renderDotPlot({
-    target,
-    dataset,
-    annotationName: annotation.name,
-    annotationValues,
-    setErrorContent: () => {},
-    setShowError: () => {},
-    genes,
-    drawCallback
-  })
-}
-
 /** Renders a Morpheus-powered dot plot for the given URL paths and annotation
   * Note that this has a lot in common with Heatmap.js.  they are separate for now
   * as their display capabilities may diverge (esp. since DotPlot is used in global gene search)
